sensorfw-core: sync the latest, Qt-less code from repowerd
I've done a port of the SocketReader code and related code to GLib APIs
in repowerd. This commit brings that code from repowerd commit
32605e47efed ("Merge branch 'focal-proper-backlight-ctrl' into 'main'")
waydroid-sensors.

This eliminates the need to link to Qt, integrates with the existing
GLib-based EventLoop better, and avoid a memory "leak" from queued
signal emission, as described in [1].

And to remove Qt from the rest of the code, also change the use of
quint64 to a standard uint64_t.

[1] https://gitlab.com/ubports/development/core/packaging/waydroid/-/merge_requests/10#note_1308000201

Fixes: https://github.com/waydroid/waydroid-sensors/issues/3